Fixes a missed uint16_t to uint32_t

+++ RotaryEncoderPCF8574.cpp
@@ -7,7 +7,7 @@ RotaryEncoderPCF8574::RotaryEncoderPCF8574(IoAbstractionRef _ioExpander, uint8_t
   clk = _clk;
 }
 
-void RotaryEncoderPCF8574::setBounds(uint16_t _lowerBound, uint16_t _upperBound) {
+void RotaryEncoderPCF8574::setBounds(uint32_t _lowerBound, uint32_t _upperBound) {
   lowerBound = _lowerBound;
   upperBound = _upperBound;  
   normalizeBounds();
@@ -26,7 +26,7 @@ void RotaryEncoderPCF8574::normalizeBounds() {
   }
 }
 
-void RotaryEncoderPCF8574::setup(uint16_t counterTimeValueFromEEPROM) {
+void RotaryEncoderPCF8574::setup(uint32_t counterTimeValueFromEEPROM) {
   ioDevicePinMode(ioExpander, sw, INPUT_PULLUP);
   ioDevicePinMode(ioExpander, dt, INPUT_PULLUP);
   ioDevicePinMode(ioExpander, clk, INPUT_PULLUP);
